fix(global): fix prerendering issue;
FIx CI Next.js prerendering issue by extracting the HomePage in its own component and importing it in app/page.tsx as dynamic with ssr:false.
