+++ frontend/src/app/hooks/useSearchHistory.ts
@@ -1,3 +1,5 @@
+"use client";
+
 import { useLocalStorage } from "@uidotdev/usehooks";
 
 export const useSearchHistory = () => {
